+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow two separate cutadapt runs to occur
 - Added a rule to generate annotation files easily with `run_annotations`
 - Added the ability to specify relative coordinates in `bwtools_query`
+- Unstranded Peak calling and motif calling
+- Modeling module with DESeq2 and stranded feature quant with HTseq-count
 
 ### Changed
 - Refactored `NETseq_pause_calling.py` for performance boost